refactor: issue #1711 - removed remotelink and changelog updated fields from jira model
